+++ src/Commands/Commands.cpp
@@ -12,6 +12,8 @@
 #include "ToggleSWSidebar.h"
 #include "FireTacticalSW.h"
 #include "ToggleMessageList.h"
+#include "BuildLastOfTab.h"
+#include "RebuildTab3Subtypes.h"
 
 #include <CCINIClass.h>
 
@@ -28,6 +30,12 @@ DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 	MakeCommand<ToggleDesignatorRangeCommandClass>();
 	MakeCommand<ToggleMessageListCommandClass>();
 	MakeCommand<ToggleSWSidebar>();
+	MakeCommand<BuildLastOfTabCommandClass<0>>();
+	MakeCommand<BuildLastOfTabCommandClass<1>>();
+	MakeCommand<BuildLastOfTabCommandClass<2>>();
+	MakeCommand<RebuildVehicleCommandClass>();
+	MakeCommand<RebuildAircraftCommandClass>();
+	MakeCommand<RebuildNavalCommandClass>();
 
 	if (Phobos::Config::SuperWeaponSidebarCommands)
 	{